Added exception handling for wrong coordinates
closes #30

If coordinates outside the specified mosaic are provided astropy will
throw a ValueException. This would lead to a cancelation of function
`write_mosaic_objects_to_pink_file_v2` and thus to a broken file since
the number of files in the header information would not equal the
actual contained number of files. Catching the exception and
decrementing the number of files for each image that can not be written
will fix this.